UI: Add QOpenGLxxx compatibility for non OpenGLES builds
Enable the OpenGL changes to be applied to all platforms with Qt >= 5.4.
To effect this edit mythrender_opengl.h and change the line
 if defined USING_OPENGLES && QT_VERSION >= QT_VERSION_CHECK(5, 4, 0)
to
 if QT_VERSION >= QT_VERSION_CHECK(5, 4, 0)

Also applies the OpenGLES only change to MythRenderOpenGL::makeCurrent
to all Qt version >= 5.4.  This prevents lots of warning messages
about swapBuffer without a previous makeCurrent.

NB using the current QGLWidget/Context calls with Qt >= 5.4, it's not
possible to use OpenGL video with the Qt UI - the video window disappears
and get lots of error messges:

QOpenGLContext::swapBuffers() called with non-exposed window, behavior is undefined

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>